// src/app/models/PackageDetails.ts
export class PackageDetails {
    id?: number;
    functions?: string;
    startDate?: Date;
    endDate?: Date;
    totalDays?: number;
    createdOn?: Date;
    competences?: string;
    shift?: string;
    packageId?: string;
    nT_User?: string;
    status?: number;
    reporterName?: string;
}