// src/app/models/Package.ts
export class Package {
    packageUID?: string;
    type?: number;
    reporterName?: string;
    department?: string;
    functions?: string;
    startDate?: Date;
    endDate?: Date;
    totalEmployees?: number;
    totalDays?: number;
    createdOn?: Date;
    competences?: string;
    shift?: string;
    nT_User?: string;
    status?: number;
}