// src/app/models/PackageDetailsHistory.ts
export class PackageDetailsHistory {
    id?: number;
    functions?: string;
    startDate?: Date;
    endDate?: Date;
    totalDays?: number;
    createdOn?: Date;
    competences?: string;
    shift?: string;
    pacakgeId?: string;
    nT_User?: string;
    status?: number;
    reporterName?: string;
    //detailsID?: number;
}