// src/app/models/PackageHistory.ts
export class PackageHistory {
    packageUID?: string;
    type?: number;
    reporterName?: string;
    department?: string;
    functions?: string;
    startDate?: Date;
    endDate?: Date;
    totalEmployees?: number;
    totalDays?: number;
    createdOn?: Date;
    competences?: string;
    shift?: string;
    nT_User?: string;
    status?: number;

}